LayoutTests/fast/dom/Element/getBoundingClientRect-vertical-td-expected.txt

1-A	1-B	1-C	1-D	1-E
2-A	2-B	2-C	2-D	2-E
3-A	3-B	3-C	3-D	3-E
4-A	4-B	4-C	4-D	4-E
5-A	5-B	5-C	5-D	5-E
PASS doesRectContainRect('1-A', gTrRect, gTdRect) is true
PASS doesRectContainRect('1-B', gTrRect, gTdRect) is true
PASS doesRectContainRect('1-C', gTrRect, gTdRect) is true
PASS doesRectContainRect('1-D', gTrRect, gTdRect) is true
PASS doesRectContainRect('1-E', gTrRect, gTdRect) is true
PASS doesRectContainRect('2-A', gTrRect, gTdRect) is true
PASS doesRectContainRect('2-B', gTrRect, gTdRect) is true
PASS doesRectContainRect('2-C', gTrRect, gTdRect) is true
PASS doesRectContainRect('2-D', gTrRect, gTdRect) is true
PASS doesRectContainRect('2-E', gTrRect, gTdRect) is true
PASS doesRectContainRect('3-A', gTrRect, gTdRect) is true
PASS doesRectContainRect('3-B', gTrRect, gTdRect) is true
PASS doesRectContainRect('3-C', gTrRect, gTdRect) is true
PASS doesRectContainRect('3-D', gTrRect, gTdRect) is true
PASS doesRectContainRect('3-E', gTrRect, gTdRect) is true
PASS doesRectContainRect('4-A', gTrRect, gTdRect) is true
PASS doesRectContainRect('4-B', gTrRect, gTdRect) is true
PASS doesRectContainRect('4-C', gTrRect, gTdRect) is true
PASS doesRectContainRect('4-D', gTrRect, gTdRect) is true
PASS doesRectContainRect('4-E', gTrRect, gTdRect) is true
PASS doesRectContainRect('5-A', gTrRect, gTdRect) is true
PASS doesRectContainRect('5-B', gTrRect, gTdRect) is true
PASS doesRectContainRect('5-C', gTrRect, gTdRect) is true
PASS doesRectContainRect('5-D', gTrRect, gTdRect) is true
PASS doesRectContainRect('5-E', gTrRect, gTdRect) is true
PASS successfullyParsed is true

TEST COMPLETE
